Fix four confirmed bugs from full app review
1. db.py: fractional shares truncated to int on load
   astype(int) silently truncated 12.5 shares to 12, causing wrong P&L
   and portfolio value for all fractional positions. Changed to float.

2. indicators.py: RSI returned NaN instead of ~100 during all-positive
   periods. avg_loss=0 produced NaN via /NaN replacement. Strong uptrends
   now correctly score RSI near 100 (avg_gain>0) or 50 (flat market).

3. risk.py: Sortino returned 0.0 when a stock had no negative excess-return
   days, making a perfect uptrend score identically to a deeply underwater
   position. Now returns 99.0 (excellent) for stocks with zero downside days
   and positive mean excess return.

4. daily_briefing.py: earnings conflict overrode other conflicts but
   downgraded severity to "Caution" (amber). A near-earnings stock with
   composite or sentiment conflicts now correctly shows "Conflicted" (red).
   Earnings-only conflict retains "Caution" when no other signals disagree.

+++ stock_analyzer/db.py
@@ -83,7 +83,7 @@ def load_holdings() -> pd.DataFrame:
             if rows:
                 df = pd.DataFrame(rows)[["ticker", "shares", "avg_cost"]]
                 df.columns = ["Ticker", "Shares", "Avg Cost ($)"]
-                df["Shares"] = df["Shares"].astype(int)
+                df["Shares"] = df["Shares"].astype(float)
                 df["Avg Cost ($)"] = df["Avg Cost ($)"].astype(float)
                 return df
             # Table is empty — return empty frame so user starts fresh
